pyproject.toml
[tool.poetry]
name = "getdata-actions-notion"
version = "0.1.0"
description = "GitHub Actionsの定期実行でデータ収集し、Notionにデータを保存するシステム"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
ccxt = "^4.1.0"
python-dotenv = "^1.0.0"
notion-client = "^2.2.0"
asyncio = "^3.4.3"
aiohttp = "^3.9.0"
pandas = "^2.1.0"
orjson = "^3.9.0"
tenacity = "^8.2.0"
pydantic = "^2.5.0"
loguru = "^0.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
isort = "^5.12.0"
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"